==> notebooks/versions/2.2-bt-deep-learning-images-RestNet50.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split
import os
import shutil
import time
import tensorflow as tf
from sklearn.metrics import accuracy_score, f1_score, confusion_matrix, classification_report
from tensorflow.keras.preprocessing.image import ImageDataGenerator, load_img, img_to_array
from tensorflow.keras.applications import ResNet50
from tensorflow.keras.applications.resnet50 import preprocess_input
from tensorflow.keras.models import Model
from tensorflow.keras.layers import Dense, GlobalAveragePooling2D
from tensorflow.keras.optimizers import Adam
from tensorflow.keras.applications.resnet50 import preprocess_input
import numpy as np
from tensorflow.keras.models import load_model
from tensorflow.keras.preprocessing import image
from tensorflow.keras.applications.resnet50 import decode_predictions

# Chargement des données et division en train, val et test  

df=pd.read_csv('X_train_update.csv')
df1=pd.read_csv("y_train_CVw08PX.csv")


#Nous allons changer les valeurs de notre colonne chemin_images du dataFrame par les chemins des images correspondantes à productid et imageid
for mmm, mmm1 in zip(df.imageid, df.productid):    
        nom_img = f"image_{mmm}_product_{mmm1}.jpg"
        df.loc[df['imageid'] == mmm, 'image_name'] = nom_img
 
#On peut supprimer les colonnes productid et imageid de notre data Frame enfin de l'alleger.
df=df.drop(["designation","productid", "imageid","description"], axis=1)

df["categ"] = df1.prdtypecode
df.head()

# Chemins
CSV_PATH = './labels.csv'
IMG_DIR = 'C:/Users/Bureau/Desktop/projetRakuten/images/train/image_train'
DEST_DIR = 'C:/Users/Bureau/Desktop/projetRakuten/images/train/dataset_split'  # dossier de destination

# Lire le CSV
#df = pd.read_csv(CSV_PATH)

# Diviser en train (70%) / val (15%) / test (15%)
df_train, df_temp = train_test_split(df, test_size=0.3, stratify=df['categ'], random_state=42)
df_val, df_test = train_test_split(df_temp, test_size=0.5, stratify=df_temp['categ'], random_state=42)

 Fonction pour copier les images dans un dossier trié
def save_split(df_split, split_name):
    for _, row in df_split.iterrows():
        try:
            label = row['categ']
            filename = row['image_name']
        except KeyError:
            print("Erreur : colonne 'label' ou 'image_name' introuvable.")
            print(f"Colonnes disponibles : {df_split.columns}")
            break  # ou continue selon ton besoin
        src_path = os.path.join(IMG_DIR, filename)
        dest_path = os.path.join(DEST_DIR, split_name, str(label))
        os.makedirs(dest_path, exist_ok=True)
        try:
            shutil.copy(src_path, os.path.join(dest_path, filename))
        except FileNotFoundError:
            print(f"Image not found: {filename}")
            
# Copier les fichiers
save_split(df_train, 'train')
save_split(df_val, 'val')
save_split(df_test, 'test')

# Création de modèle 

# Chemins
base_dir = 'C:/Users/Bureau/Desktop/projetRakuten/images/train/dataset_split'
img_size = (224, 224)
batch_size = 32

# Générateurs de données
train_gen = ImageDataGenerator(preprocessing_function=preprocess_input)
val_gen = ImageDataGenerator(preprocessing_function=preprocess_input)
test_gen = ImageDataGenerator(preprocessing_function=preprocess_input)

train_data = train_gen.flow_from_directory(
    os.path.join(base_dir, 'train'),
    target_size=img_size,
    batch_size=batch_size,
    class_mode='categorical'
)

val_data = val_gen.flow_from_directory(
    os.path.join(base_dir, 'val'),
    target_size=img_size,
    batch_size=batch_size,
    class_mode='categorical'
)

test_data = test_gen.flow_from_directory(
    os.path.join(base_dir, 'test'),
    target_size=img_size,
    batch_size=batch_size,
    class_mode='categorical',
    shuffle=False
)


# Charger ResNet50 sans le top
base_model = ResNet50(weights='imagenet', include_top=False, input_shape=(224, 224, 3))

# Ne pas entraîner les couches du modèle de base
for layer in base_model.layers:
    layer.trainable = False

# Ajouter les couches de classification
x = base_model.output
x = GlobalAveragePooling2D()(x)
x = Dense(512, activation='relu')(x)
predictions = Dense(train_data.num_classes, activation='softmax')(x)

model = Model(inputs=base_model.input, outputs=predictions)

# Compiler
model.compile(optimizer=Adam(learning_rate=0.0001),
              loss='categorical_crossentropy',
              metrics=['accuracy'])



# Entraînement
model.fit(
    train_data,
    validation_data=val_data,
    epochs=20
)

# Prédictions
y_true = test_data.classes  
y_pred_probs = model.predict(test_data)
y_pred = np.argmax(y_pred_probs, axis=1)

# Évaluation sur le test
loss, acc = model.evaluate(test_data)
print(f"Test accuracy: {acc:.2%}")
# F1-score macro
f1 = f1_score(y_true, y_pred, average='macro')
print(f"F1-score (macro) sur les données de validation : {f1:.4f}")

# Sauvegarde du modèle
model.save('resnet50_rakuten1.h5')


## Calcul du F1-score moyen (macro, micro ou weighted selon le besoin)

MODEL_PATH = 'resnet50_rakuten1.h5'
model1 = load_model(MODEL_PATH)

# Prédictions
y_true = test_data.classes  # les vraies classes (entiers)
y_pred_probs = model1.predict(test_data)
y_pred = np.argmax(y_pred_probs, axis=1)

# Évaluation sur le test
loss, acc = model1.evaluate(test_data)
print(f"Test accuracy: {acc:.2%}")
# F1-score :
for f1_sc in ["macro", "micro", "weighted"]:
    f1 = f1_score(y_true, y_pred, average=f1_sc)  
    print(f"F1-score moyen {f1_sc}:", f1)
